備份

+++ CraftDailyCorner/DTOs/SortUpdateDTO.cs
@@ -2,7 +2,7 @@
 {
     public class SortUpdateDTO
     {
-        public int ItemId { get; set; }
+        public long ItemId { get; set; }
         public byte SortOrder { get; set; }
     }
 }
