+++ CraftDailyCorner/ViewModels/CreatorPortfolio/Front/VMPortfolioDetailItem.cs
@@ -2,7 +2,7 @@
 {
     public class VMPortfolioDetailItem
     {
-        public int ItemID { get; set; }
+        public long ItemID { get; set; }
         public string ImageUrl { get; set; } = null!;
     }
 }
