+++ CraftDailyCorner/ViewModels/CreatorPortfolio/VMCreatorPortfolioItemEdit.cs
@@ -1,6 +1,6 @@
 ﻿public class VMCreatorPortfolioItemEdit
 {
-    public int ItemID { get; set; }
+    public long ItemID { get; set; }
 
     public string ImageUrl { get; set; } = null!;
